Create LoginController
